src/requirements.txt
pillow
customtkinter
tk
ctkmessagebox
